move __proxy_call__ to test
